Stop using Metrics.NET reporter loop
Due to the transient nature of PerformanceCounters, Metrics.NET
register once, nrever unregister model is not quite an ideal fit
for the PerfCounterReporter.

We now rescan the files before each report, and update the set
of PerfCounters we are watching. Once we have the set we are
reporting for this round, we report. We also clean up any
unused PerfCounter objects at this point.